requirements/base.txt
Django==1.11.16
  pytz==2018.5
psycopg2==2.7.5
django-dotenv==1.4.1
huey==1.10.3
redis==2.10.6
pip==18.1
